fix modularity

+++ CommunityDetectionAlgorithm/louvain-code/graph.cpp
@@ -139,7 +139,7 @@ Graph::display_binary(char *filename, char *filename_w, int type) {
   foutput.write((char *)(&s),4);
 
   // outputs cumulative degree sequence
-  long tot=0;
+  long long tot=0;
   for (unsigned int i=0 ; i<s ; i++) {
     tot+=(long)links[i].size();
     foutput.write((char *)(&tot),8);
